requirements.txt
opencv-contrib-python
numpy
matplotlib
tensorflow
keras
python-socketio
Flask
eventlet
